# Improve partitions performance